Export NSEventMask and add everything under addon namespace
