+++ test/npm-debug.log
@@ -1,0 +1,23 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/local/Cellar/node/6.6.0/bin/node',
+1 verbose cli   '/usr/local/bin/npm',
+1 verbose cli   'run',
+1 verbose cli   'build' ]
+2 info using npm@4.0.2
+3 info using node@v6.6.0
+4 verbose stack Error: ENOENT: no such file or directory, open '/Users/pronebird/freelance/mullvad/nseventmonitor/test/package.json'
+4 verbose stack     at Error (native)
+5 verbose cwd /Users/pronebird/freelance/mullvad/nseventmonitor/test
+6 error Darwin 16.4.0
+7 error argv "/usr/local/Cellar/node/6.6.0/bin/node" "/usr/local/bin/npm" "run" "build"
+8 error node v6.6.0
+9 error npm  v4.0.2
+10 error path /Users/pronebird/freelance/mullvad/nseventmonitor/test/package.json
+11 error code ENOENT
+12 error errno -2
+13 error syscall open
+14 error enoent ENOENT: no such file or directory, open '/Users/pronebird/freelance/mullvad/nseventmonitor/test/package.json'
+15 error enoent ENOENT: no such file or directory, open '/Users/pronebird/freelance/mullvad/nseventmonitor/test/package.json'
+15 error enoent This is most likely not a problem with npm itself
+15 error enoent and is related to npm not being able to find a file.
+16 verbose exit [ -2, true ]
